wakeupword/train/requirements.txt
numpy
scipy
PyYAML
piper-tts
tqdm
datasets
torch
torchaudio
torchcodec
coqui-tts
